pyproject.toml
[tool.poetry]
name = "ai-agent-system"
version = "0.1.0"
description = "Production-ready multi-agent AI system with FastAPI and async orchestration"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "ai_agent_system"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.0"
uvicorn = {version = "^0.23.2", extras = ["standard"]}
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
httpx = {version = "^0.25.0", extras = ["http2"]}
aioredis = "^2.0.1"
redis = "^5.0.0"
SQLAlchemy = {version = "^2.0.20", extras = ["asyncio"]}
asyncpg = "^0.29.0"
alembic = "^1.12.0"
tenacity = "^8.2.3"
structlog = "^23.1.0"
loguru = "^0.7.0"
python-json-logger = "^2.0.7"
celery = "^5.3.4"
kombu = "^5.3.2"
flower = "^2.0.1"
prometheus-client = "^0.17.1"
opentelemetry-api = "^1.20.0"
opentelemetry-sdk = "^1.20.0"
opentelemetry-instrumentation-fastapi = "^0.41b0"
opentelemetry-exporter-otlp = "^1.20.0"
cryptography = "^41.0.3"
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
python-jose = "^3.3.0"
pyjwt = "^2.8.0"
email-validator = "^2.0.0.post2"
cachetools = "^5.3.1"
apscheduler = "^3.10.4"
openai = "^1.3.5"
anthropic = "^0.3.11"
beautifulsoup4 = "^4.12.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.9.1"
flake8 = "^6.1.0"
mypy = "^1.5.1"
requests-mock = "^1.11.0"
responses = "^0.24.1"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
check_untyped_defs = true
disallow_any_unimported = true
disallow_any_expr = true
disallow_any_decorated = true
disallow_any_explicit = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_ignores = true
strict_equality = true
plugins = []

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-q --disable-warnings --maxfail=1"
asyncio_mode = "auto"

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
per-file-ignores = {"__init__.py" = ["F401"]}

[tool.isort]
profile = "black"
line_length = 100
